correccion de error

+++ js/Banco.js
@@ -2,13 +2,9 @@ let sumaSaldosAcreedores = 0;
 let numeroCuenta, nombreCliente, saldoActual, estadoCuenta;
 
 while (numeroCuenta < 0) {
-  numeroCuenta = parseInt(
-    prompt(
-      "Ingrese el número de cuenta (ingrese un valor negativo para finalizar:"
-    )
-  );
+  numeroCuenta = parseInt(prompt("Ingrese el número de cuenta (ingrese un valor negativo para finalizar:"));
   nombreCliente = prompt("Ingrese el nombre del cliente:");
-  saldoActual = parseFloat(prompt("Ingrese el saldo actual:"));
+  saldoActual = parseInt(prompt("Ingrese el saldo actual:"));
 
   if (saldoActual > 0) {
     estadoCuenta = "Acreedor";
